requirements.txt
torch >= 1.12.1
torchvision >= 0.13.0
numpy >= 1.19.5
pillow~=10.0.1
setuptools~=60.2.0
tqdm~=4.62.3
gdown~=4.7.1
PyYAML~=6.0.1
requests~=2.31.0
matplotlib~=3.8.0
scipy~=1.11.3
scikit-image~=0.22.0
loguru~=0.7.2
zenodo-get~=1.5.1
graphviz~=0.20.1
opencv-python-headless~=4.8.1.78 # Using headless version for compatibility with graphviz
captum~=0.7.0
Augmentor~=0.2.12
mkdocs
pydoc-markdown
pymdown-extensions
python-markdown-math
pandas~=2.2.1